add some basic vertex code
