+++ .gitignore
@@ -1,3 +1,6 @@
+# Mac
+.DS_Store
+
 # Python-generated files
 __pycache__/
 *.py[oc]
