+++ app/backends/dependencies.py
@@ -1,14 +1,13 @@
-from typing import Union
 from fastapi import HTTPException
 from app.config.settings import BACKEND_MAP
 from app.schema.open_ai import ChatCompletionRequest
-from app.schema.open_ai import CohereRequest
+from app.schema.open_ai import EmbeddingRequest
 
 class Backend:
     def __init__(self, capability: str):
         self.capability = capability
 
-    def __call__(self, req: Union[ChatCompletionRequest, CohereRequest]):
+    def __call__(self, req: ChatCompletionRequest | EmbeddingRequest):
         model_id = req.model
 
         backend, capability = BACKEND_MAP.get(model_id, (None, None))
