+++ app/routers/api_v1.py
@@ -3,8 +3,7 @@ from fastapi import APIRouter, Depends
 from app.auth.dependencies import RequiresScope
 from app.backends.dependencies import Backend
 from app.auth.schemas import Scope
-from app.schema.open_ai import ChatCompletionRequest, ChatCompletionResponse
-from app.schema.open_ai import CohereRequest
+from app.schema.open_ai import ChatCompletionRequest, ChatCompletionResponse, EmbeddingRequest
 
 router = APIRouter()
 
@@ -19,7 +18,7 @@ async def converse(
 
 @router.post("/embeddings")
 async def embeddings(
-    req: CohereRequest,
+    req: EmbeddingRequest,
     api_key=Depends(RequiresScope([Scope.MODELS_EMBEDDING])),
     backend=Depends(Backend('embedding'))
 ):
